+++ ChangeLog.md
@@ -21,12 +21,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ChangeLog added
 ### Changed
 - __BREAKING:__ sql files changed, update from 0.1.0 to 0.2.0 not possible, __fresh install required__
+  - lims_results.sql: unit varchar(10) added
+  - limx_methods.sql: fk_product changed to fk_equipment
 - Samples: Client sampling person is now one of the options empty-yes-no
 - Methods: At creation (New Method) 'Status' is not visible
 - Limits: At creation (New Limit Set) 'Status' is not visible
 - Results: 'Abnormality' is relabeld to 'Nonconformity'
 - Test Report: Statement of Conformity based on the information whether all results are within the respective measurement range of the method.
-- lims_results.sql: unit varchar(10) added
 ### Removed
 - Samples: At creation the option to create a contact for client sample taker 
 - Samples: Proposal selection is now hidden.